feat: move personalTokens from Account field to Individual field
